feat: add support for circular references in the swagger file
Close #60

BREAKING CHANGE:

swagger-mock-validator now supports validating schemas containing circular references. Prior to this change schemas containing circular references were ignored. If you have a schema that contains circular references and a mock that is not compatible with that schema the swagger-mock-validator will now fail.

It is recommended that consumers and providers coordinate upgrading to this release so that both sides agree on what is considered valid vs invalid.