feat: removed swagger warnings from validation results
BREAKING CHANGE:
Prior to this change swagger-mock-validator would return any pact/swagger warnings or errors as part of the validation result. Now when a validation error is encountered with a swagger or pact file the tool will fail with an error. Warnings are no longer reported at all. The list of validation result codes returned by the programatic API have been updated to reflect this change.