fix: do not fail if the spec defines unsupported security requirements
BREAKING CHANGE:
Prior to this change the tool used to fail when a pact request with no apparent auth was validated against a swagger operation containing security requirements both supported and unsupported by the tool (e.g. basic auth and OAuth2). Now if at least one unsupported security requirement is defined in the swagger operation the validation will pass. It is recommended that consumers and providers coordinate upgrading to this release so that both sides agree on what is considered valid vs invalid.