fix: do not fail if an OpenApi 3 spec defines an empty security requirement
BREAKING CHANGE:
Prior to this change the tool used to fail when a pact request with no auth was validated against an OpenApi 3 operation containing a security requirement (e.g. basic auth) plus an empty security requirement (which is normally used to represent that auth is optional). Now empty security requirements are considered for OpenApi3 specs hence not failing when the request lacks auth attributes.