fix: correctly validate request bodies containing a JSON null or false value
BREAKING CHANGE: Prior to this change the tool used to succeed when a pact request specified a request body containing solely the JSON literals ‘null' or 'false’, regardless of what the spec accepted. Now, request bodies with such values will only pass validation if the spec is okay with them.